requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
spotipy==2.23.0
python-dotenv==1.0.0
httpx>=0.24,<0.26
Pillow==10.2.0
